Extract backend analytics contract reporting
